+++ app/core/config.py
@@ -8,13 +8,12 @@ class Settings(BaseSettings):
     SQLALCHEMY_DATABASE_URI: str = "sqlite:///./test.db"
 
     # Version
-    BACKEND_VERSION: str = "2.5.0"
+    BACKEND_VERSION: str = "2.6.0"
 
     # Security
     SECRET_KEY: str = "a_very_secret_key"
     ALGORITHM: str = "HS256"
     ACCESS_TOKEN_EXPIRE_MINUTES: int = 30
-    API_KEY: str = "a_super_secret_api_key"
 
     # General
     TIMEZONE: str = "Asia/Shanghai"